feat(profile-edit): hydrate industries, interests and expertises from SSR (#239)
Split edit/page.tsx (previously 'use client') into a server page that
fetches industries, interests, and expertises via the SSR fetchers
(24h revalidate), and a client container that holds all the existing
form / hooks / dynamic-import logic and now receives the three lists
as initialData. useLocations stays per-user for now (separate ticket).
